Reorganize `reference` and `shared`
