+++ packages/seroval/src/tree/constants.ts
@@ -1,5 +1,5 @@
+import { SerovalConstant, SerovalNodeType } from '../constants';
 import type { SerovalConstantNode } from './types';
-import { SerovalConstant, SerovalNodeType } from './types';
 
 function createConstantNode(value: SerovalConstant): SerovalConstantNode {
   return {
