+++ packages/seroval/src/tree/serialize.ts
@@ -9,8 +9,8 @@ import {
   markRef,
 } from './context';
 import { serializeConstant } from './constants';
-import { GLOBAL_KEY } from './reference';
-import { isValidIdentifier } from './shared';
+import { GLOBAL_KEY } from '../reference';
+import { isValidIdentifier } from '../shared';
 import { SYMBOL_STRING } from '../symbols';
 import type {
   SerovalAggregateErrorNode,
